pyproject.toml
[project]
name = "sql-mcp"
version = "0.1.0"
dependencies = [
  "fastmcp",
  "sqlalchemy",
  "psycopg2-binary",
  "sqlalchemy-redshift",
  "redshift_connector",
  "sqlparse",
  "python-dotenv",
  "pyyaml"
]